bd-1716: Endre datatype for beløp i overgangsstønad fra Int til BigDecimal

+++ src/main/kotlin/no/nav/bidrag/transport/behandling/inntekt/request/TransformerInntekterRequest.kt
@@ -56,5 +56,5 @@ data class Overgangsstonad(
     val periodeTil: LocalDate?,
 
     @Schema(description = "Beløp overgangsstønad")
-    val belop: Int
+    val belop: BigDecimal
 )
